requirements.txt
absl-py==0.9.0
alabaster==0.7.12
astor==0.8.1
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.1.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
efficientnet @ git+https://github.com/qubvel/efficientnet@e9fdd43857785fe5ccf3863915dcaf618b86849f
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.8.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
helpdev==0.7.1
idna==2.9
image-classifiers @ git+https://github.com/qubvel/classification_models.git@a0f006e05485a34ccf871c421279864b0ccd220b
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.3.0
ipython
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.2
joblib
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring
kiwisolver
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.5
numpydoc==1.0.0
opencv-python==4.2.0.34
openpyxl
openslide-python
opt-einsum==3.2.1
packaging==20.4
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
rope==0.17.0
scikit-image
scikit-learn
scipy==1.4.1
SecretStorage==3.1.2
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
yapf==0.30.0
zipp
